MSVC Warning Fix (v140_xp) - 7 size_t to int results in narrowing? ------ stddef.h: #define offsetof(s,m) ((size_t)&reinterpret_cast<char const volatile&>((((s*)0)->m)))

+++ rehlds/engine/delta.cpp
@@ -39,7 +39,7 @@ typedef struct delta_link_s
 typedef struct delta_definition_s
 {
 	char *fieldName;
-	int fieldOffset;
+	size_t fieldOffset;
 } delta_definition_t;
 
 typedef struct delta_definition_list_s
